Forberedelse til å kopiere over EØS-skjema: Flytting av kode (#2526)
* Trekk opprettelse av personopplysningsgrunnlag ut i egen service

* Trekk opprettelse av vilkårsvurdering ut i egen service

* Rett skrivefeil

* Rename variabel forrigePersongrunnlagBarna -> barnMedTilkjentYtelseIForrigeBehandling

* Send med forrigeBehandlingSomErVedtatt i stedet for å hente på nytt

* Flytt funksjoner som kun ble kalt fra VilkårsvurderingForNyBehandlingService inn i den filen

* (glemt i tidligere commit) Send med forrigeBehandlingSomErVedtatt

* Rett opp tester som ikke hadde fått oppdatert hvilken service funksjon ligger i

* Trekk ut tester i egen fil som tester funksjoner i VilkårsvurderingForNyBehandlingService

* Legg til AbstractSpringIntegrationTest på ny test

* Bytt ut kall til personopplysningGrunnlagRepository med service-kall i Vilkårservice

* Gjør forrigeBehandlingSomErVedtatt obligatorisk ved årsak endre migreringsdato

* Kast feil i stedet for å bruke !! på nyMigreringsdato for migreringsbehandlinger

* Trekk funksjonell kode ut i utils-fil

* Trekk funksjonell kode ut i utils-fil for helmanuell migrering
