+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/steg/VilkårsvurderingForNyBehandlingService.kt
@@ -1,0 +1,113 @@
+package no.nav.familie.ba.sak.kjerne.steg
+
+import no.nav.familie.ba.sak.common.Feil
+import no.nav.familie.ba.sak.kjerne.behandling.BehandlingService
+import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
+import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
+import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersongrunnlagService
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårService
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingService
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkårsvurdering
+import org.slf4j.LoggerFactory
+import org.springframework.stereotype.Service
+import java.time.LocalDate
+
+@Service
+class VilkårsvurderingForNyBehandlingService(
+    private val vilkårService: VilkårService,
+    private val vilkårsvurderingService: VilkårsvurderingService,
+    private val behandlingService: BehandlingService,
+    private val persongrunnlagService: PersongrunnlagService
+) {
+
+    fun opprettVilkårsvurderingUtenomHovedflyt(
+        behandling: Behandling,
+        forrigeBehandlingSomErVedtatt: Behandling?,
+        nyMigreringsdato: LocalDate? = null
+    ) {
+        when (behandling.opprettetÅrsak) {
+            BehandlingÅrsak.ENDRE_MIGRERINGSDATO -> {
+                genererVilkårsvurderingForMigreringsbehandlingMedÅrsakEndreMigreringsdato(
+                    behandling = behandling,
+                    forrigeBehandlingSomErVedtatt = forrigeBehandlingSomErVedtatt
+                        ?: throw Feil("Kan ikke opprette behandling med årsak 'Endre migreringsdato' hvis det ikke finnes en tidligere behandling som er iverksatt"),
+                    nyMigreringsdato = nyMigreringsdato
+                        ?: throw Feil("Kan ikke opprette behandling med årsak 'Endre migreringsdato' uten en migreringsdato")
+                )
+                // Lagre ned migreringsdato
+                behandlingService.lagreNedMigreringsdato(nyMigreringsdato, behandling)
+            }
+            BehandlingÅrsak.HELMANUELL_MIGRERING -> {
+                genererVilkårsvurderingForHelmanuellMigrering(
+                    behandling = behandling,
+                    nyMigreringsdato = nyMigreringsdato
+                        ?: throw Feil("Kan ikke opprette behandling med årsak 'Helmanuell migrering' uten en migreringsdato")
+                )
+                // Lagre ned migreringsdato
+                behandlingService.lagreNedMigreringsdato(nyMigreringsdato, behandling)
+            }
+            !in listOf(BehandlingÅrsak.SØKNAD, BehandlingÅrsak.FØDSELSHENDELSE) -> {
+                vilkårService.initierVilkårsvurderingForBehandling(
+                    behandling = behandling,
+                    bekreftEndringerViaFrontend = true,
+                    forrigeBehandlingSomErVedtatt = forrigeBehandlingSomErVedtatt
+                )
+            }
+            else -> logger.info(
+                "Perioder i vilkårsvurdering generer ikke automatisk for " +
+                    behandling.opprettetÅrsak.visningsnavn
+            )
+        }
+    }
+
+    fun genererVilkårsvurderingForMigreringsbehandlingMedÅrsakEndreMigreringsdato(
+        behandling: Behandling,
+        forrigeBehandlingSomErVedtatt: Behandling,
+        nyMigreringsdato: LocalDate
+    ): Vilkårsvurdering {
+
+        val vilkårsvurdering = Vilkårsvurdering(behandling = behandling).apply {
+            personResultater =
+                VilkårsvurderingForNyBehandlingUtils.lagPersonResultaterForMigreringsbehandlingMedÅrsakEndreMigreringsdato(
+                    vilkårsvurdering = this,
+                    nyMigreringsdato = nyMigreringsdato,
+                    forrigeBehandlingVilkårsvurdering = hentForrigeBehandlingVilkårsvurdering(
+                        forrigeBehandlingSomErVedtatt = forrigeBehandlingSomErVedtatt,
+                        nyBehandlingId = behandling.id
+                    ),
+                    personopplysningGrunnlag = persongrunnlagService.hentAktivThrows(behandling.id)
+                )
+        }
+        return vilkårsvurderingService.lagreNyOgDeaktiverGammel(vilkårsvurdering = vilkårsvurdering)
+    }
+
+    private fun hentForrigeBehandlingVilkårsvurdering(
+        forrigeBehandlingSomErVedtatt: Behandling,
+        nyBehandlingId: Long
+    ) = (
+        vilkårsvurderingService
+            .hentAktivForBehandling(forrigeBehandlingSomErVedtatt.id)
+            ?: throw Feil(
+                "Kan ikke kopiere vilkårsvurdering fra forrige behandling ${forrigeBehandlingSomErVedtatt.id}" +
+                    "til behandling $nyBehandlingId"
+            )
+        )
+
+    fun genererVilkårsvurderingForHelmanuellMigrering(
+        behandling: Behandling,
+        nyMigreringsdato: LocalDate
+    ): Vilkårsvurdering {
+        val vilkårsvurdering = Vilkårsvurdering(behandling = behandling).apply {
+            personResultater = VilkårsvurderingForNyBehandlingUtils.lagPersonResultaterForHelmanuellMigrering(
+                vilkårsvurdering = this,
+                nyMigreringsdato = nyMigreringsdato,
+                personopplysningGrunnlag = persongrunnlagService.hentAktivThrows(behandling.id)
+            )
+        }
+        return vilkårsvurderingService.lagreNyOgDeaktiverGammel(vilkårsvurdering = vilkårsvurdering)
+    }
+
+    companion object {
+        private val logger = LoggerFactory.getLogger(this::class.java)
+    }
+}
