+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/steg/VilkårsvurderingForNyBehandlingUtils.kt
@@ -1,0 +1,111 @@
+package no.nav.familie.ba.sak.kjerne.steg
+
+import no.nav.familie.ba.sak.kjerne.autovedtak.fødselshendelse.Resultat
+import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlag
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingMigreringUtils
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.PersonResultat
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkår
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.VilkårResultat
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkårsvurdering
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.gjelderAlltidFraBarnetsFødselsdato
+import java.time.LocalDate
+
+object VilkårsvurderingForNyBehandlingUtils {
+
+    fun lagPersonResultaterForMigreringsbehandlingMedÅrsakEndreMigreringsdato(
+        vilkårsvurdering: Vilkårsvurdering,
+        forrigeBehandlingVilkårsvurdering: Vilkårsvurdering,
+        nyMigreringsdato: LocalDate,
+        personopplysningGrunnlag: PersonopplysningGrunnlag
+    ): Set<PersonResultat> {
+        return personopplysningGrunnlag.søkerOgBarn.map { person ->
+            val personResultat = PersonResultat(vilkårsvurdering = vilkårsvurdering, aktør = person.aktør)
+
+            val vilkårTyperForPerson = forrigeBehandlingVilkårsvurdering.personResultater
+                .single { it.aktør == person.aktør }.vilkårResultater
+                .filter { it.resultat == Resultat.OPPFYLT }
+                .map { it.vilkårType }
+
+            val vilkårResultater = vilkårTyperForPerson.map { vilkår ->
+                val fom = VilkårsvurderingMigreringUtils.utledPeriodeFom(
+                    forrigeBehandlingVilkårsvurdering,
+                    vilkår,
+                    person,
+                    nyMigreringsdato
+                )
+
+                val tom: LocalDate? =
+                    VilkårsvurderingMigreringUtils.utledPeriodeTom(
+                        forrigeBehandlingVilkårsvurdering,
+                        vilkår,
+                        person,
+                        fom
+                    )
+
+                val begrunnelse = "Migrering"
+
+                VilkårResultat(
+                    personResultat = personResultat,
+                    erAutomatiskVurdert = false,
+                    resultat = Resultat.OPPFYLT,
+                    vilkårType = vilkår,
+                    periodeFom = fom,
+                    periodeTom = tom,
+                    begrunnelse = begrunnelse,
+                    behandlingId = personResultat.vilkårsvurdering.behandling.id
+                )
+            }.toSortedSet(VilkårResultat.VilkårResultatComparator)
+
+            val manglendePerioder = VilkårsvurderingMigreringUtils.kopiManglendePerioderFraForrigeVilkårsvurdering(
+                vilkårResultater,
+                forrigeBehandlingVilkårsvurdering, person
+            )
+            vilkårResultater.addAll(manglendePerioder.map { it.kopierMedParent(personResultat) }.toSet())
+            personResultat.setSortedVilkårResultater(vilkårResultater)
+
+            personResultat
+        }.toSet()
+    }
+
+    fun lagPersonResultaterForHelmanuellMigrering(
+        vilkårsvurdering: Vilkårsvurdering,
+        nyMigreringsdato: LocalDate,
+        personopplysningGrunnlag: PersonopplysningGrunnlag
+    ): Set<PersonResultat> {
+        return personopplysningGrunnlag.søkerOgBarn.map { person ->
+            val personResultat = PersonResultat(vilkårsvurdering = vilkårsvurdering, aktør = person.aktør)
+
+            val vilkårTyperForPerson = Vilkår.hentVilkårFor(person.type)
+            val vilkårResultater = vilkårTyperForPerson.map { vilkår ->
+                val fom = when {
+                    vilkår.gjelderAlltidFraBarnetsFødselsdato() -> person.fødselsdato
+                    nyMigreringsdato.isBefore(person.fødselsdato) -> person.fødselsdato
+                    else -> nyMigreringsdato
+                }
+
+                val tom: LocalDate? = when (vilkår) {
+                    Vilkår.UNDER_18_ÅR -> person.fødselsdato.plusYears(18)
+                        .minusDays(1)
+                    else -> null
+                }
+
+                val begrunnelse = "Migrering"
+
+                VilkårResultat(
+                    personResultat = personResultat,
+                    erAutomatiskVurdert = false,
+                    resultat = Resultat.OPPFYLT,
+                    vilkårType = vilkår,
+                    periodeFom = fom,
+                    periodeTom = tom,
+                    begrunnelse = begrunnelse,
+                    behandlingId = personResultat.vilkårsvurdering.behandling.id
+                )
+            }.toSortedSet(VilkårResultat.VilkårResultatComparator)
+
+            personResultat.setSortedVilkårResultater(vilkårResultater)
+
+            personResultat
+        }.toSet()
+    }
+}
