+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vilkårsvurdering/VilkårService.kt
@@ -22,13 +22,11 @@ import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ba.sak.kjerne.endretutbetaling.EndretUtbetalingAndelService
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.Person
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonType
+import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersongrunnlagService
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlag
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlagRepository
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import no.nav.familie.ba.sak.kjerne.personident.PersonidentService
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingMigreringUtils.kopiManglendePerioderFraForrigeVilkårsvurdering
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingMigreringUtils.utledPeriodeFom
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingMigreringUtils.utledPeriodeTom
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingUtils.flyttResultaterTilInitielt
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingUtils.lagFjernAdvarsel
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingUtils.muterPersonResultatDelete
@@ -56,6 +54,7 @@ class VilkårService(
     private val personidentService: PersonidentService,
     private val featureToggleService: FeatureToggleService,
     private val endretUtbetalingAndelService: EndretUtbetalingAndelService,
+    private val persongrunnlagService: PersongrunnlagService
 ) {
 
     fun hentVilkårsvurdering(behandlingId: Long): Vilkårsvurdering? = vilkårsvurderingService.hentAktivForBehandling(
@@ -191,7 +190,7 @@ class VilkårService(
             )
         }
 
-        val personopplysningGrunnlag = hentAktivPersonopplysningGrunnlagThrows(behandling.id)
+        val personopplysningGrunnlag = persongrunnlagService.hentAktivThrows(behandling.id)
         if (personopplysningGrunnlag.søkerOgBarn
             .single { it.aktør == personidentService.hentAktør(restNyttVilkår.personIdent) }.type != PersonType.SØKER
         ) {
@@ -211,7 +210,7 @@ class VilkårService(
         bekreftEndringerViaFrontend: Boolean,
         forrigeBehandlingSomErVedtatt: Behandling? = null
     ): Vilkårsvurdering {
-        val personopplysningGrunnlag = hentAktivPersonopplysningGrunnlagThrows(behandling.id)
+        val personopplysningGrunnlag = persongrunnlagService.hentAktivThrows(behandling.id)
 
         if (behandling.skalBehandlesAutomatisk && personopplysningGrunnlag.barna.isEmpty()) {
             throw IllegalStateException("PersonopplysningGrunnlag for fødselshendelse skal inneholde minst ett barn")
@@ -309,33 +308,8 @@ class VilkårService(
         }
     }
 
-    fun genererVilkårsvurderingForMigreringsbehandlingMedÅrsakEndreMigreringsdato(
-        behandling: Behandling,
-        forrigeBehandlingSomErVedtatt: Behandling? = null,
-        nyMigreringsdato: LocalDate
-    ): Vilkårsvurdering {
-        val vilkårsvurdering = Vilkårsvurdering(behandling = behandling).apply {
-            personResultater = lagVilkårsvurderingForMigreringsbehandlingMedÅrsakEndreMigreringsdato(
-                this,
-                forrigeBehandlingSomErVedtatt!!,
-                nyMigreringsdato
-            )
-        }
-        return vilkårsvurderingService.lagreNyOgDeaktiverGammel(vilkårsvurdering = vilkårsvurdering)
-    }
-
-    fun genererVilkårsvurderingForHelmanuellMigrering(
-        behandling: Behandling,
-        nyMigreringsdato: LocalDate
-    ): Vilkårsvurdering {
-        val vilkårsvurdering = Vilkårsvurdering(behandling = behandling).apply {
-            personResultater = lagVilkårsvurderingForHelmanuellMigrering(this, nyMigreringsdato)
-        }
-        return vilkårsvurderingService.lagreNyOgDeaktiverGammel(vilkårsvurdering = vilkårsvurdering)
-    }
-
     private fun lagTomVilkårsvurdering(vilkårsvurdering: Vilkårsvurdering): Set<PersonResultat> {
-        val personopplysningGrunnlag = hentAktivPersonopplysningGrunnlagThrows(vilkårsvurdering.behandling.id)
+        val personopplysningGrunnlag = persongrunnlagService.hentAktivThrows(vilkårsvurdering.behandling.id)
 
         return personopplysningGrunnlag.søkerOgBarn.map { person ->
             val personResultat = PersonResultat(vilkårsvurdering = vilkårsvurdering, aktør = person.aktør)
@@ -360,7 +334,7 @@ class VilkårService(
     }
 
     private fun lagManuellVilkårsvurdering(vilkårsvurdering: Vilkårsvurdering): Set<PersonResultat> {
-        val personopplysningGrunnlag = hentAktivPersonopplysningGrunnlagThrows(vilkårsvurdering.behandling.id)
+        val personopplysningGrunnlag = persongrunnlagService.hentAktivThrows(vilkårsvurdering.behandling.id)
 
         return personopplysningGrunnlag.søkerOgBarn.map { person ->
             genererPersonResultatForPerson(vilkårsvurdering, person)
@@ -373,7 +347,7 @@ class VilkårService(
     ): Set<PersonResultat> {
         val barnaAktørSomAlleredeErVurdert = personidentService.hentAktørIder(barnaSomAlleredeErVurdert)
 
-        val personopplysningGrunnlag = hentAktivPersonopplysningGrunnlagThrows(vilkårsvurdering.behandling.id)
+        val personopplysningGrunnlag = persongrunnlagService.hentAktivThrows(vilkårsvurdering.behandling.id)
         val annenForelder = personopplysningGrunnlag.annenForelder
         val eldsteBarnSomVurderesSinFødselsdato =
             personopplysningGrunnlag.barna.filter { !barnaAktørSomAlleredeErVurdert.contains(it.aktør) }
@@ -436,7 +410,7 @@ class VilkårService(
     }
 
     private fun lagVilkårsvurderingForMigreringsbehandling(vilkårsvurdering: Vilkårsvurdering): Set<PersonResultat> {
-        val personopplysningGrunnlag = hentAktivPersonopplysningGrunnlagThrows(vilkårsvurdering.behandling.id)
+        val personopplysningGrunnlag = persongrunnlagService.hentAktivThrows(vilkårsvurdering.behandling.id)
 
         return personopplysningGrunnlag.søkerOgBarn.map { person ->
             val personResultat = PersonResultat(vilkårsvurdering = vilkårsvurdering, aktør = person.aktør)
@@ -454,101 +428,6 @@ class VilkårService(
 
                 val tom: LocalDate? =
                     if (vilkår == Vilkår.UNDER_18_ÅR) person.fødselsdato.plusYears(18).minusDays(1) else null
-
-                val begrunnelse = "Migrering"
-
-                VilkårResultat(
-                    personResultat = personResultat,
-                    erAutomatiskVurdert = false,
-                    resultat = Resultat.OPPFYLT,
-                    vilkårType = vilkår,
-                    periodeFom = fom,
-                    periodeTom = tom,
-                    begrunnelse = begrunnelse,
-                    behandlingId = personResultat.vilkårsvurdering.behandling.id
-                )
-            }.toSortedSet(VilkårResultatComparator)
-
-            personResultat.setSortedVilkårResultater(vilkårResultater)
-
-            personResultat
-        }.toSet()
-    }
-
-    private fun lagVilkårsvurderingForMigreringsbehandlingMedÅrsakEndreMigreringsdato(
-        vilkårsvurdering: Vilkårsvurdering,
-        forrigeBehandlingSomErVedtatt: Behandling,
-        nyMigreringsdato: LocalDate
-    ): Set<PersonResultat> {
-        val forrigeBehandlingsvilkårsvurdering = vilkårsvurderingService
-            .hentAktivForBehandling(forrigeBehandlingSomErVedtatt.id)
-            ?: throw Feil(
-                "Kan ikke kopiere vilkårsvurdering fra forrige behandling ${forrigeBehandlingSomErVedtatt.id}" +
-                    "til behandling ${vilkårsvurdering.behandling.id}"
-            )
-        val personopplysningGrunnlag = hentAktivPersonopplysningGrunnlagThrows(vilkårsvurdering.behandling.id)
-
-        return personopplysningGrunnlag.søkerOgBarn.map { person ->
-            val personResultat = PersonResultat(vilkårsvurdering = vilkårsvurdering, aktør = person.aktør)
-
-            val vilkårTyperForPerson = forrigeBehandlingsvilkårsvurdering.personResultater
-                .single { it.aktør == person.aktør }.vilkårResultater
-                .filter { it.resultat == Resultat.OPPFYLT }
-                .map { it.vilkårType }
-
-            val vilkårResultater = vilkårTyperForPerson.map { vilkår ->
-                val fom = utledPeriodeFom(forrigeBehandlingsvilkårsvurdering, vilkår, person, nyMigreringsdato)
-
-                val tom: LocalDate? =
-                    utledPeriodeTom(forrigeBehandlingsvilkårsvurdering, vilkår, person, fom)
-
-                val begrunnelse = "Migrering"
-
-                VilkårResultat(
-                    personResultat = personResultat,
-                    erAutomatiskVurdert = false,
-                    resultat = Resultat.OPPFYLT,
-                    vilkårType = vilkår,
-                    periodeFom = fom,
-                    periodeTom = tom,
-                    begrunnelse = begrunnelse,
-                    behandlingId = personResultat.vilkårsvurdering.behandling.id
-                )
-            }.toSortedSet(VilkårResultatComparator)
-
-            val manglendePerioder = kopiManglendePerioderFraForrigeVilkårsvurdering(
-                vilkårResultater,
-                forrigeBehandlingsvilkårsvurdering, person
-            )
-            vilkårResultater.addAll(manglendePerioder.map { it.kopierMedParent(personResultat) }.toSet())
-            personResultat.setSortedVilkårResultater(vilkårResultater)
-
-            personResultat
-        }.toSet()
-    }
-
-    private fun lagVilkårsvurderingForHelmanuellMigrering(
-        vilkårsvurdering: Vilkårsvurdering,
-        nyMigreringsdato: LocalDate
-    ): Set<PersonResultat> {
-        val personopplysningGrunnlag = hentAktivPersonopplysningGrunnlagThrows(vilkårsvurdering.behandling.id)
-
-        return personopplysningGrunnlag.søkerOgBarn.map { person ->
-            val personResultat = PersonResultat(vilkårsvurdering = vilkårsvurdering, aktør = person.aktør)
-
-            val vilkårTyperForPerson = Vilkår.hentVilkårFor(person.type)
-            val vilkårResultater = vilkårTyperForPerson.map { vilkår ->
-                val fom = when {
-                    vilkår.gjelderAlltidFraBarnetsFødselsdato() -> person.fødselsdato
-                    nyMigreringsdato.isBefore(person.fødselsdato) -> person.fødselsdato
-                    else -> nyMigreringsdato
-                }
-
-                val tom: LocalDate? = when (vilkår) {
-                    Vilkår.UNDER_18_ÅR -> person.fødselsdato.plusYears(18)
-                        .minusDays(1)
-                    else -> null
-                }
 
                 val begrunnelse = "Migrering"
 
@@ -588,11 +467,6 @@ class VilkårService(
                 .vilkårResultater.any { it.vilkårType == Vilkår.UTVIDET_BARNETRYGD }
         }
         return false
-    }
-
-    private fun hentAktivPersonopplysningGrunnlagThrows(behandlingId: Long): PersonopplysningGrunnlag {
-        return personopplysningGrunnlagRepository.findByBehandlingAndAktiv(behandlingId)
-            ?: throw IllegalStateException("Fant ikke personopplysninggrunnlag for behandling $behandlingId")
     }
 
     private fun finnPersonResultatForPersonThrows(
